bug fix
the plan was so that if someone has over 5 points, then the player had to manually restart the game using a button. The Problem was that the game could be played further in the background. I tried to fix ist but failed, so I removed the replay button and the game would now autoamtically replay after someone got 5 Points